jedec: Do not return an error when no chip is found
Returning an error code other than QIPROG_SUCCESS has interesting
consequences. The endpoint will be stalled and the control transaction
aborted. Instead of getting a QIPROG_ID_INVALID, the host gets an error.
Both these cases indicate that a chip was not identified, but the stalled
endpoint situation could also indicate an error with the VultureProg
device.

Since it is possible to respond to the read_chip_id request even when no chip
is found, always respond with QIPROG_ID_INVALID instead of an error code.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>